Added Gaussian function
Used the exponential class as a template to add the Gaussian function.
Guess only works for positive peaks.

Right now, y0 is just the minimum value in values.

Guessing y0 can be tricky. As an option, it can average all points that
are further away than 3 standard deviations from the average of the
values. If it doesn't find any points that far away from the mean, it
returns y0=0. In this case, there are either not enough points to get a
good baseline (in which case the guess will be ok) or the amplitude is
very very low (in which case it may or may not be a good guess,
depending on the offset).

Not sure which method is better long-term.