Added spectral Delay Correction Curves
I'm using the same kind of curve object and file format for spectral
delay corrections. This way, we can use all the same framework for
tuning curves and spectral delay correction.

I'm curious to see if we can correlate the mixer positions and spectral
delay correction values. They aren't in the same file as of now.