Adding 2D Gaussian Fits to wt.fit
I'm pretty sure Fitter DOES NOT work with 2D fit functions, but you can
fit any slice if you give it the data values and axis values. There is
relatively robust support for different shapes and lengths for the
values and axis arrays, but they do have to make sense.