MAINT: Use register_projection to enable WT Axes
Thanks to @anntzer for the suggestion, it worked perfectly (with a
couple of lines needed to actually properly register the class)

This addresses half of the deprecation warnings found in #890

This also enables using our axes class without calling our own `Figure`
class.

```
plt.subplot(111, projection="wright")
plt.plot(data)
```

Question for @untzag/@darienmorrow, Do we wish to document the above
behavior and make the projection name more public facing than it
already is (currently considered implementation detail, especially as it
was not accessible by other means previously).
If so, do we also wish to name it "wright_tools" instead (or some other
variation) to make it more explicitly linked to our library?