Update _join.py
Added error message for multidimensional axes. This checks all data objects. I thought that would be of use given that one may have to define 1D versions of variable axes for all objects, and could make a mistake in one but not all of the 'datas' to be joined. Open to discussion.